Add -host-whitelist option
